*: Use uint64_t for weight down the path to Zebra

+++ bgpd/bgp_zebra.h
@@ -136,7 +136,7 @@ extern void bgp_zebra_update_fib_install_pending(struct bgp_dest *dest,
 						 struct bgp *bgp, bool install);
 
 struct in6_addr *bgp_path_info_to_ipv6_nexthop(struct bgp_path_info *path, ifindex_t *ifindex);
-bool bgp_zebra_use_nhop_weighted(struct bgp *bgp, struct attr *attr, uint32_t *nh_weight);
+bool bgp_zebra_use_nhop_weighted(struct bgp *bgp, struct attr *attr, uint64_t *nh_weight);
 
 bool update_ipv4nh_for_route_install(int nh_othervrf, struct bgp *nh_bgp, struct in_addr *nexthop,
 				     struct attr *attr, bool is_evpn, struct zapi_nexthop *api_nh);
